pyproject.toml
[tool.poetry]
name = "flask-api"
version = "0.1.0"
description = ""
authors = ["Esteban Rodriguez <https://github.com/esteban-alo>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.1"
flask = "^3.0.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.25"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.2.1"
black = "^23.12.1"
coverage = "^7.4.0"
isort = "^5.13.2"
mypy = "^1.8.0"
pylint = "^3.0.3"
pytest = "^7.4.4"
ruff = "^0.1.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
